Created predict model, updated README file
>
