+++ README.md
@@ -79,12 +79,60 @@ The GET `/api/year` endpoint supports the following query parameters:
 
 #### Prediction Endpoints
 
-| Method | Endpoint                            | Description                             |
-| ------ | ----------------------------------- | --------------------------------------- |
-| POST   | `/api/predict/carbon_emissions/:id` | Returns an array of cars, limited to 20 |
-| POST   | `/api/predict/carbon_emissions2`    | Returns an array of cars, limited to 20 |
-| POST   | `/api/predict/price/:id`            | Returns a car with a specific ID        |
+| Method | Endpoint                            | Body             | Description                                      |
+| ------ | ----------------------------------- | ---------------- | ------------------------------------------------ |
+| POST   | `/api/predict/carbon_emissions/:id` | None             | Returns a prediction for CO2 generated           |
+| POST   | `/api/predict/price/:id`            | None             | Returns a prediction for the price of the car    |
+| POST   | `/api/predict/carbon_emissions2`    | Array of car ids | Returns an array of CO2 predictions, limit of 10 |
 
+Format of POST `/api/predict/carbon_emissions/:id`:
+
+```
+{
+  "id": 1,
+  "make": "Tesla",
+  "model": "X",
+  "year": 2018,
+  "predicted_carbon_emissions": 3213.0
+}
+```
+
+Format of POST `/api/predict/price/:id`:
+
+```
+{
+  "id": 1,
+  "make": "Tesla",
+  "model": "X",
+  "year": 2018,
+  "predicted_price": 26000
+}
+```
+
+`/api/predict/carbon_emissions2` must be called with a body that has an array of car ids to get predictions for. It will then return an array of predictions.
+
+Input:
+```
+[1, 3123154123]
+```
+
+```
+[
+  {
+    "id": 1,
+    "make": "Tesla",
+    "model": "X",
+    "year": 2018,
+    "predicted_price": 26000
+    "status": 200,
+  },
+  {
+    "id": 3123154123,
+    "error": "Could not find car with id",
+    "status": 404
+  }
+]
+```
 # Data Model
 
 See `epa_vehicles_schema.txt`.
